+++ pubspec.yaml
@@ -29,3 +29,6 @@ dependencies:
 
 flutter:
   uses-material-design: true
+
+  assets:
+    - assets/info.txt
